Give back the alpha the author wrote, by two rules for two things written
`csscolorparser` holds its channels as `f32`, so `rgba(0, 0, 0, 0.1)` reached
a caller of the addon's `parseColor` as 0.10000000149011612 -- an internal
float width leaking through a public boundary. v1 answers 0.102, quantising
alpha to eight bits; the browser answers 0.1, and the browser is the tiebreak
here as it is for the mix clamp.

Channels are now presented as the shortest decimal that identifies the `f32`,
which for anything a person types is what they typed. That covers decimals and
percentages -- `33%` lands on the same `f32` as `0.33` and prints the same --
and it does not cover a hex byte: `#0000007f` is `127/255`, and the shortest
decimal naming that `f32` is 0.49803922, which is neither the byte nor the
ratio. So the hex branch computes `byte / 255` where the byte is known. Two
rules, because the authors wrote two different things.

The scaling to 0..255 stays in `f32`, where `#808080` is exactly 128. Widening
first and multiplying in `f64` lands a byte a hair under itself, which would
have broken rows that already passed -- a test pins that, since the trap is in
the fix rather than in the defect.

`parse_channels` returns `[f64; 4]` and `extended_srgb` parses at `f64` from
the text, so `color(srgb ...)` keeps the author's number rather than the
nearest `f32` to it. The addon drops its now-useless widening conversion.

`animate.test.ts`'s alpha test goes from `it.fails` to `it`: it failed on both
surfaces and passes on both now, which is the evidence the fix reached the
shared parser rather than one side of it. Two more tests in that file --
`mixColor` and `mix` over colours -- were failing at HEAD and pass here.

Also closes the trait's other instantiation: `Sampled` was exercised only for
`f64`, so the uniformity it states was proven for numbers and assumed for
colours. A colour crossfade now goes through a track, a plan and a group, and
`total_duration` on `Sampled` records why the count is a `usize` where
JavaScript takes a fraction.

+++ crates/meo-canvas-node/src/lib.rs
@@ -205,7 +205,7 @@ fn parse_color(mut cx: FunctionContext<'_>) -> JsResult<'_, JsValue> {
     let object = cx.empty_object();
     for (name, channel) in [("r", red), ("g", green), ("b", blue), ("a", alpha)]
     {
-        let value = cx.number(f64::from(channel));
+        let value = cx.number(channel);
         object.set(&mut cx, name, value)?;
     }
     Ok(object.upcast())
